Fix LP_TITLE_CLOSE on bash inside screen
As the TITLE_CLOSE sequence for screen contains a '\', and this char is
the escape char for PS1 in bash we have to escape it.
Cc: @pkkolos

Note: there are more places where % or \ should be more properly
escaped, but this is a huge work. This is a bit like XSS in HTML.
In angel-PS1 I solved the issue by design.