zsh: fix LP_OLD_PROMPT_COMMAND (by not using it at all!)
In bash only the $PROMPT_COMMAND can be used for hooks and we
have to wrap the previous hook with our own.
zsh has the ${precmd_functions[*]} array for multiple prompt hook
functions, so the user is free to install any hooks he want along
liquidprompt. Also, we don't replace anymore the main `precmd` function
(as we did on ealier versions of liquidprompt) so we don't have to wrap
it either.

So let's drop that kind of bash emulation on zsh.