Move contributors to a separate file
Contributors were tracked in the CHANGES file for each release, as well
as being tracked in liquidprompt itself, which not only cluttered the
file, but most of the contributors weren't listed at all.
This new file should list ALL contributors, both code and Github issue
kind. (This again took forever).
The going forward plan is to add contributors to this file, and also to
call out specific contributors with each release notes in the release
tag.

Update README to point to new CONTRIBUTORS file.
Tweak the .mailmap file to remove pointless duplicate lines.