Local specify IFS in each case it is used
Liquidprompt was relying on the default IFS for most needs. This was
somewhat safe, as it saved the IFS value from startup, and used that
when it needed an IFS value. The problem cropped up when the user set a
blank, weird, or completely unset value for IFS before sourcing
Liquidprompt. In this case some `read` calls would fail.

The simple fix is anywhere IFS is used, to make a local IFS with the
exact value that is needed. There are only two cases that use IFS:
* `read`, which uses IFS to split the string into fields to put into the
  specified variables.
* Unquoted $var or $(command) references, often passed into `for` or
  similar. Note that this doesn't apply to unquoted variable assignments
  (`foo=$bar`, doesn't need to be quoted, will never split).

Fixes #613