Call prompt_on() when lp_activate() is run
This will be the same outcome for the normal usage: source liquidprompt.
But for the reloading of the user's config, if they would only run
lp_activate(), some things would not be reset correctly. Namely,
directory location specific things, like the current working directory.
This is the simplest fix. It also means that we can put any other
resetting code in prompt_on() to keep it central.