Fix shorten path breaking on config reload
Since config reloading is supported now with only a run of
lp_activate(), any unsetting of functions will break Liquidprompt if the
config option is changed. This was the case with __lp_shorten_path(), as
it was unset.

To fix, wrap __lp_shorten_path() with a __lp_path() function, or wrap
__lp_set_dirtrim() if that was enabled instead. This way we can keep
both in memory for when the config is changed.

Also move out the theme specific stuff to the _lp_theme_directory()
function. Now that $lp_path is set every time we change directory,
either with the shell symbol for the path or the literal path string
(possibly shortened), we can just use that.