Add __lp_theme_list to list themes loaded in memory
All this does is scan for _lp_<theme>_theme_prompt() functions, and
print themes matching that. Also add a --list parameter to lp_theme()
that calls this function.

Add bash and zsh completion functions as well, since this is the only
"public" user function that takes parameters, might as well complete for
it. I often found myself pressing TAB often while working on this, so I
thought it's probably a good idea.