Breakout hostname determine code from color
This wasn't done before because I thought that deciding whether to
display the hostname or not would be up to the themes. But I now think
that it is more useful to make the hostname display work the same
accross all themes, and obey the LP_HOSTNAME_ALWAYS config option. This
is what _lp_hostname() now does, and now _lp_hostname_color() only
colors the hostname based on the connection, and adds chroot and a "@"
to it.