Make LP_COLOR_MARK_* options go to LP_COLOR_MARK
Since these are global options, they should just work with any theme,
not just the default.

It was kinda like this before, and I changed it, thinking that all color
options would be only for the default theme, but I think I've changed my
mind.

This greatly simplifies _lp_username_color(), and cuts down on duplicate
_lp_user() calls.