Add caching for git diff --shortstat calls
Processing of git diff --shortstat outputs was already duplicated
between uncommitted() and unstaged() functions. This breakout helps fix
that, since the output format of --shorstat does not change when the
targets of the diff change.

By saving the output of the --shortstat in a cache, calls to both the
files() and lines() functions do not duplicate a git call. This does not
happen in the default theme, but I'm sure some theme will want to do it.
And my testing shows that this change is no slower for it.