Fix binary file changes not showing VCS as dirty
This was introduced in the VCS rework. As binary files do not have
"lines" as most VCS providers would understand them, changes to binary
files do not report as different lines in a diff. This is why the real
check for a dirty repository is if any (tracked) files have been
changed.

This "revert" adds back calls that previously existed, so for all VCS
providers, it is no slower than it was in v1.12.1.

In the case for Git, since 8bf1772d771904bd2095a974b35795c9db2c96cc
lines() and files() share a cache, and therefore would have no speed
loss. But since Git supports staging, the _lp_vcs_uncommitted_lines()
call will never happen, so it is no change for Git.

Fixes #647