Fix new Git diff cache never filling on Zsh
If this code is run on Bash, Zsh, and Ksh93:
    function test_func {
      typeset foobar
      [[ -z ${foobar+x} ]]
      echo $?
    }; test_func

Bash and Ksh return 0, while Zsh returns 1.

According to man zshbuiltins:
> Except as noted below for control flags that change the behavior, a
> parameter is created for each name that does not already refer to one.
> When inside a function, a new parameter is created for every name
> (even those that already exist), and is unset again when the function
> completes. See `Local Parameters' in zsh‐param(1).

The metioned section in zsh‐param(1) does not give any more hints.

So it would seem that when declaring a variable as local in Zsh, it sets
it to the empty string.

As far as I can tell with the relevant POSIX standard:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_02
There is no specification as to how these tests act when the variable is
a local. And as for the typeset command:

> If the command name [is typeset], the results are unspecified.

So the solution is that these caches can only be treated as invalid if
they are empty or unset, not just unset.

The ${var-} form (note the "-") needs to be used for the shells that do
not treat declaring a local as setting it when the shell is run in
'errorunset' mode.

The most annoying thing about this lesson is that I had already learned
it and made a note of it in tests/test_shell.sh line 158. Embarrassing.