Complete truncate_chars_from_path_left feature
In Liquidprompt v1, this path truncation was simple; just cut out the
correct number of characters. But now that we can have a "special"
directory in the middle of the path (the VCS root) and also each
separator and directory needs its own formatting, this is much harder.

The extra tricky part is that a check is needed to see if we are
currently shortening already, and if so, not to fail out of the shorten
if it looks like we would "break even".