Fix unset errors on Bash 3.2
Bash 3.2 scans a whole line for unset problems, unlike later versions,
which will exit the line if the boolean expression returns False
already.

This SO post is super helpful: https://stackoverflow.com/a/61551944/8338372
This is why the ${array[@]+"${array[@]}"} form is used, anywhere either
the array could be empty, or the array could have empty strings as
values. Most of our array usages we know for sure there are good values
in, but user config ones we have no way to know for sure.