Remove backslash escaping from Zsh __lp_escape()
This was added in commit 2d53a4a1c07ed4e8fd3634f8feeb0e09875386e6. I
don't understand why, except that maybe a single backslash display would
look like an escape code for a control character, like \n. But that is
not the case for Bash, so why the difference?

Zsh only uses percents (%) to escape things in the prompt, where Bash
only uses bashslashes. This changes makes the same input string print
the same way in the prompt for both shells.