Fix local bug in Zsh-5.0
Zsh 5.0 continues to be a huge pain. As noted in the shell wiki, all
variables assigned to another variable as part of a 'local' or 'typeset'
line must be quoted, or things will break. The reason I caught these
many instances was a failing test on Zsh-5.0.8 because of the bad line
in __lp_pwd_tilde().