Disable shellcheck warnings in theme-preview.sh
SC2329 is function never invoked. Since the functions are invoked in the
themes, all possible data functions need to be defined to support any
possible themes.

SC2317 is command appears to be unreachable. This must have been a false
positive previously, and in newer Shellcheck versions (0.11.0) it no
longer is.